requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
alembic==1.13.0
openai==1.3.0
anthropic==0.7.0
rouge-score==0.1.2
nltk==3.8.1
pandas==2.1.3
numpy==1.26.2
python-dotenv==1.0.0
pytest==7.4.3
httpx==0.25.1
jinja2==3.1.2
plotly==5.18.0
python-multipart==0.0.6
aiofiles==23.2.1
redis==5.0.1
celery==5.3.4
